misc/HtmlTestSuite/results/test-0408.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<link href="mailto:obrufau@igalia.com" rel="author" title="Oriol Brufau"/>
<link href="mailto:mstensho@chromium.org" rel="author" title="Morten Stenshorne"/>
<link href="https://drafts.csswg.org/css-scroll-anchoring/#candidate-examination" rel="help"/>
<meta content="The candidate examination algorithm iterates abspos descendants of the containing block. Being inside a fragmentainer shouldn't break that." name="assert"/>
<style>
html {
  column-count: 1; /* Fragmentainer */
}

main {
  position: relative; /* Containing block */
}

div {
  position: absolute; /* Abspos */
  font-size: 100px;
  width: 200px;
  height: 4000px;
  line-height: 100px;
}
</style>
</head><body><main><div>abc <b id="b">def</b> ghi</div></main>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
// Tests anchoring to a text node that is moved by preceding text,
// everything in an absolutely positioned element whose containing block
// is inside a multi-column fragmentainer.

test(() =&gt; {
  var b = document.querySelector("#b");
  var preText = b.previousSibling;
  document.scrollingElement.scrollTop = 150;
  preText.nodeValue = "abcd efg ";
  assert_equals(document.scrollingElement.scrollTop, 250);
}, "Anchoring with text wrapping changes.");
</script>
</body></html>
